requirements.txt
ray==2.9.3
yfinance==0.2.36
pandas==2.2.1
click==8.1.7
kubernetes==29.0.0
python-dotenv==1.0.1
requests==2.31.0

# Development dependencies
pytest==8.0.0
flake8==7.0.0
mypy==1.8.0
black==24.1.1
isort==5.13.2 